feat: split tagging a release version and creating Github release
By default new git tags are only created by go-semantic-release when
there is a conventional commit that fits the default behaviour. This
means that `chore` does not bump to a new patch version. More info on
this can be found
[here](https://github.com/go-semantic-release/commit-analyzer-cz#how-the-commit-messages-are-analyzed).

This splits the tagging of a new release from creating a new Github
release.

This allows for new Github releases to only be created when there is
a new `feat` or `fix`.

An alternative would be to use a custom commit analyser but then extend
would be diverging from the conventional commits spec.