fix: update release workflow
This reverts the splitting of tagging and running goreleaser as done in
[#131](https://github.com/einride/extend/pull/131) and instead replaces
the flow so that go-semantic-release will call goreleaser as a hook when
there is a new version to release so that goreleaser can then updload
the other assets.

I'm not 100% sure this will achieve the desired effect of being able to
not fail when we get commits that shouldn't create a new release but
will merge, test and iterate.